# Swagger/swagger.yml
openapi: 3.0.4
info:
  title: Domino
  description: |
    API para contar puntos en juego de domino
  # termsOfService: https://swagger.io/terms/
  contact:
    email: jgenao588@gmail.com
  # license:
  #   name: Apache 2.0
  #   url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
# externalDocs:
#   description: Find out more about Swagger
#   url: https://swagger.io
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /sessions:
    post:
      tags:
        - Sessions
      summary: Create new session sending.
      description: Create a new session to your device seding a device id.
      operationId: createSession
      requestBody:
        description: Create your new session
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSession'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    CreateSession:
      required:
        - deviceId
      type: object
      properties:
        # id:
        #   type: integer
        #   format: int64
        #   example: 10
        deviceId:
          type: string
          example: uuid
        # category:
        #   $ref: '#/components/schemas/Category'
        # photoUrls:
        #   type: array
        #   xml:
        #     wrapped: true
        #   items:
        #     type: string
        #     xml:
        #       name: photoUrl
        # tags:
        #   type: array
        #   xml:
        #     wrapped: true
          # items:
          #   $ref: '#/components/schemas/Tag'
        # status:
        #   type: string
        #   description: pet status in the store
        #   enum:
        #     - available
        #     - pending
        #     - sold
      xml:
        name: createSession
...
    SessionCreated:
      type: object
      properties:
        sessionId: 
          type: string
          example: sessionId
      xml:
        name: sessionCreated
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...